# perf(http): remove allocations checking upgrade and connection header values